requirements.txt
pyrogram
tgcrypto
countryinfo
pytz
motor
aiofiles
dnspython
python-dotenv
